+++ module/metanthropes/combat.mjs
@@ -178,6 +178,14 @@ export class MetanthropesCombat extends Combat {
 		if (!this.started)
 			return ui.notifications.warn("You must begin the encounter before progressing to the next Round!");
 		await super.nextRound();
+		//* Active Effect Expiration
+		//? Fetch Active Effects with a duration of 'None' and toggle them off
+		let expiredEffects = [];
+		for (let combatant of this.combatants.values()) {
+			let effects = combatant.actor.effects.filter((e) => e.duration.label === "None");
+			expiredEffects.push(...effects);
+		}
+		await Promise.all(expiredEffects.map(e => e.update({ disabled: true })));
 		//* End of Round Effects
 		//? Iterate over Combatants
 		for (let combatant of this.combatants.values()) {
@@ -305,39 +313,35 @@ export class MetanthropesCombat extends Combat {
 		//* Update the Cycle and Round values
 		//? Get the most recent Cycle and Round values from the Combat document
 		let cycle = (await this.getFlag("metanthropes-system", "cycle")) || 1;
-		let cycleRound = (await this.getFlag("metanthropes-system", "cycleRound")) || 1;
 		switch (this.round) {
 			case 1:
 				cycle = 1;
-				cycleRound = 1;
 				break;
 			case 2:
 				cycle = 1;
-				cycleRound = 2;
 				break;
 			case 3:
 				cycle = 2;
-				cycleRound = 1;
 				break;
 			default:
 				if (this.round > 2 && (this.round - 1) % 2 === 0) {
 					cycle++;
-					cycleRound = 1;
-				} else {
-					cycleRound = 2;
 				}
 				break;
 		}
+		//? Set cycle as part of Combat document
 		this.cycle = cycle;
-		this.cycleRound = cycleRound;
+		//? Set cycle as a flag in the Combat document
 		await this.setFlag("metanthropes-system", "cycle", cycle);
-		await this.setFlag("metanthropes-system", "cycleRound", cycleRound);
+		if (e.duration.label === "None") {
+			e.update({ disabled: !e.disabled });
+		}
 		//? Reroll initiative for all combatants at the start of a new Cycle
-		if (cycle > 1 && cycleRound === 1) {
+		if (cycle > 1 && cycle % 2 !== 0) {
 			await this.resetAll();
 			this.setupTurns();
 			await ChatMessage.create({
-				content: `<br>New Cycle: ${cycle} Round: ${cycleRound}<br><br>Roll Inititiative!<br><br>`,
+				content: `<br>New Cycle: ${cycle} Round: ${this.round}<br><br>Roll Inititiative!<br><br>`,
 				speaker: {
 					alias: "Metanthropes Combat",
 				},
@@ -345,7 +349,7 @@ export class MetanthropesCombat extends Combat {
 		} else {
 			//? Create a chat message indicating the new Round
 			await ChatMessage.create({
-				content: `<br>Cycle: ${cycle} Round: ${cycleRound}<br><br>`,
+				content: `<br>Cycle: ${cycle} Round: ${this.round}<br><br>`,
 				speaker: {
 					alias: "Metanthropes Combat",
 				},